fix: proxy-serviceを実装注入パターンに変更
動的インポートがCI環境で正しく動作しなかったため、
実装注入パターンに変更:

- RewriteRuleProxyService.tsからcontainer.tsへの依存を完全に削除
- setRewriteRuleProxyServiceImpl()でbackground.tsから実装を注入
- background.tsで実装を作成してから登録

これによりContent Scriptでモジュールがロードされても
Background用のcontainerは読み込まれない
