+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/RewriteRuleProxyService.ts
@@ -1,7 +1,5 @@
 import { defineProxyService } from '@webext-core/proxy-service';
 
-import type { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
-
 /**
  * RewriteRule取得用のプロキシサービスインターフェース
  * Content Script から Background へのルール取得に使用
@@ -21,30 +19,37 @@ export interface IRewriteRuleProxyService {
 }
 
 /**
+ * サービス実装の格納場所
+ * background.ts から setRewriteRuleProxyServiceImpl() で注入される
+ */
+let serviceImpl: IRewriteRuleProxyService | null = null;
+
+/**
+ * サービス実装を注入する
+ * background.ts の main() 内で registerRewriteRuleProxyService() の前に呼び出す
+ *
+ * @param impl サービス実装
+ */
+export function setRewriteRuleProxyServiceImpl(impl: IRewriteRuleProxyService): void {
+  serviceImpl = impl;
+}
+
+/**
  * @webext-core/proxy-service を使用したRewriteRuleプロキシサービス
  * Background Script で実行され、Content Script からのルール取得を仲介
  *
- * 重要: container.ts は動的インポートを使用
- * これにより Content Script でこのモジュールがロードされても
- * Background 用のコンテナは読み込まれない
+ * 重要: このファイルは container.ts に依存しない
+ * 実装は background.ts から setRewriteRuleProxyServiceImpl() で注入される
+ * これにより Content Script でこのモジュールがロードされても問題が発生しない
  */
 function createRewriteRuleProxyService(): IRewriteRuleProxyService {
-  return {
-    async getAllRules() {
-      // 動的インポートでcontainerをロード（Background側でのみ実行される）
-      const { container } = await import('src/frameworks-and-drivers/di/container');
-      const repository = container.resolve<IRewriteRuleRepository>('IRewriteRuleRepository');
-      const rules = await repository.getAll();
-
-      return rules.toArray().map((rule) => ({
-        id: rule.id,
-        oldString: rule.oldString,
-        newString: rule.newString,
-        urlPattern: rule.urlPattern,
-        isRegex: rule.isRegex,
-      }));
-    },
-  };
+  if (!serviceImpl) {
+    throw new Error(
+      'RewriteRuleProxyService implementation not set. ' +
+      'Call setRewriteRuleProxyServiceImpl() before registerRewriteRuleProxyService().'
+    );
+  }
+  return serviceImpl;
 }
 
 export const [registerRewriteRuleProxyService, getRewriteRuleProxyService] =
